v0/requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-WTF>=1.1
WTForms>=3.0
SQLAlchemy>=2.0
python-dotenv>=1.0
